requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
jsonschema==4.20.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Security & Encryption
cryptography==41.0.7
secrets==1.0.0
hashlib==1.0.0

# Async & Networking
aiohttp==3.9.1
aiofiles==23.2.1
httpx==0.25.2
requests==2.31.0
urllib3==2.1.0

# GitHub Integration
pygithub==2.1.1
gitpython==3.1.40

# Security Scanners (External Tools)
# Note: These tools need to be installed separately:
# - semgrep: pip install semgrep
# - trufflehog: see https://github.com/trufflesecurity/trufflehog
# - osv-scanner: see https://github.com/google/osv-scanner
# - cdxgen: npm install -g @cyclonedx/cdxgen
# - syft: see https://github.com/anchore/syft

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Logging & Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing (dev dependencies)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality (dev dependencies)
black==23.12.1
flake8==6.1.0
mypy==1.7.1
isort==5.13.2
bandit==1.7.5

# Documentation (dev dependencies)
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Utilities
click==8.1.7
rich==13.7.0
tabulate==0.9.0
tqdm==4.66.1